+++ tests/sansio/test_utils.py
@@ -1,7 +1,10 @@
+from __future__ import annotations
+
 import typing as t
 
 import pytest
 
+from werkzeug.sansio.utils import get_content_length
 from werkzeug.sansio.utils import get_host
 
 
@@ -30,3 +33,23 @@ def test_get_host(
     expected: str,
 ) -> None:
     assert get_host(scheme, host_header, server) == expected
+
+
+@pytest.mark.parametrize(
+    ("http_content_length", "http_transfer_encoding", "expected"),
+    [
+        ("2", None, 2),
+        (" 2", None, 2),
+        ("2 ", None, 2),
+        (None, None, None),
+        (None, "chunked", None),
+        ("a", None, 0),
+        ("-2", None, 0),
+    ],
+)
+def test_get_content_length(
+    http_content_length: str | None,
+    http_transfer_encoding: str | None,
+    expected: int | None,
+) -> None:
+    assert get_content_length(http_content_length, http_transfer_encoding) == expected