+++ CHANGES.rst
@@ -1,11 +1,24 @@
 .. currentmodule:: werkzeug
 
-Version 2.3.7
+Version 2.3.8
 -------------
 
 Unreleased
 
+
+Version 2.3.7
+-------------
+
+Released 2023-08-14
+
 -   Use ``flit_core`` instead of ``setuptools`` as build backend.
+-   Fix parsing of multipart bodies. :issue:`2734`
+    Adjust index of last newline in data start. :issue:`2761`
+-   ``_plain_int`` and ``_plain_float`` strip whitespace before type
+    enforcement. :issue:`2734`
+-   Fix empty file streaming when testing. :issue:`2740`
+-   Clearer error message when URL rule does not start with slash. :pr:`2750`
+-   ``Accept`` ``q`` value can be a float without a decimal part. :issue:`2751`
 
 
 Version 2.3.6